fix(test): repair the performance and benchmark build tags, then guard them
Both tagged suites were broken on main and had been for an unknown length of
time. Nothing reported it because no CI lane set either tag — and absence of a
signal reads exactly like success.

The two failures had different causes:

**`performance`** — `stress_test.go` called `ensureTestBucket`/`cleanupTestObjects`,
defined only in `real_aws_integration_test.go` behind `//go:build integration`.
The two tags are never set together, so the caller could never see the callee.
Moved the shared helpers to `real_aws_helpers_test.go` tagged
`integration || performance`, which is satisfiable from either suite. The bucket
NAMES stay per-suite — a shared name is unused under whichever tag doesn't
reference it, which trips staticcheck.

**`benchmark`** — `benchmark_test.go:598` called `config.S3Client.Options()`, but
`S3Client` is an `interface{}` field and has no methods. The concrete `*s3.Client`
was already in scope as a local; read `Options()` off that instead.

Then the guard, because repairing the files fixes today and not tomorrow:
`scripts/ci/check-build-tags.sh` runs `go vet -tags <t> ./...` for every custom
tag, wired into test.yml as "Build Tags Compile". The tag list is DERIVED from
`//go:build` lines rather than hardcoded, so a new tag is covered the moment it
appears — a hardcoded list would rot the same way these files did. It found a
third tag the issue didn't mention: `benchmarks` (plural, in tests/benchmarks/dvc),
distinct from `benchmark` and also unbuilt by CI. It compiles.

These suites need real AWS credentials to RUN, so the check does not run them.
Compiling is the part that was regressing.

Verified the guard fails against the pre-fix tree: stashing the two fixes makes
it report both original errors verbatim and exit 1. staticcheck's finding count
is unchanged at 15 (all pre-existing).

This is the tag analogue of the #308/#311/#325 duplicate-drift pattern: code
nothing compiles stops tracking reality.

Closes #329
