+++ .github/workflows/test.yml
@@ -17,6 +17,31 @@ jobs:
   # identical `go test ./... -short` plus a coverage profile — a separate "Test"
   # job would be pure duplication, so it's intentionally omitted.
 
+  build-tags:
+    name: Build Tags Compile
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
+
+      - name: Set up Go
+        uses: actions/setup-go@924ae3a1cded613372ab5595356fb5720e22ba16 # v6.5.0
+        with:
+          go-version: ${{ env.GO_VERSION }}
+          cache: true
+
+      - name: Download dependencies
+        run: go mod download
+
+      # Compile-checks every custom build tag in the tree. The `performance` and
+      # `benchmark` suites need real AWS to RUN, so nothing here runs them — but
+      # nothing compiled them either, and both had been broken on main for an
+      # unknown length of time (#329). Compiling is the part that regresses.
+      - name: Check all build tags compile
+        run: bash scripts/ci/check-build-tags.sh
+
   e2e-quickstart:
     name: Quick Start E2E (emulator)
     runs-on: ubuntu-latest
